2019/2/19 by Boyin

+++ identiFace/src/main/java/com/example/dao/CustomerDAO.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import com.example.entity.*;
 
 @RepositoryRestResource(collectionResourceRel = "Customer", path = "Customer")
-public interface CustomerDAO extends CrudRepository<Customer, Integer> {
+public interface CustomerDAO extends CrudRepository<Customer, Long> {
 
 }
